src/requirements.txt
numpy==1.15.4
tqdm==4.28.1
matplotlib==3.0.2
opencv-python-3.4.5.20
matplotlib==3.0.2
scipy==1.2.0
scikit-image==0.14.1
scikit-learn==0.20.2
Pillow==5.3.0
unittest2==1.1.0
networkx==2.2